requirements.txt
Flask
pymongo
pika
twilio
python-dotenv
flasgger 